Old_Code/requirements.txt
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.12.0
astor==0.8.1
attrdict==2.0.1
babel==2.17.0
bce-python-sdk==0.9.57
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
contourpy==1.3.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.3
decorator==5.2.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
filelock==3.20.2
fire==0.7.1
Flask==3.1.2
flask-babel==4.0.0
fonttools==4.61.1
fsspec==2025.12.0
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
modelscope==1.33.0
more-itertools==10.8.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
paddleocr==2.7.0.3
paddlepaddle-xpu @ https://paddle-whl.bj.bcebos.com/paddlex/xpu/paddlepaddle_xpu-2.6.1-cp310-cp310-linux_aarch64.whl#sha256=c2bce365201cc674e49e56dbc34b5474fb8ef91249e464b1533f6d14ec8ee4ae
pandas==2.3.3
pdf2docx==0.5.8
pillow==12.1.0
premailer==3.10.0
prettytable==3.17.0
protobuf==6.33.2
psutil==7.2.1
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.20.2
pyparsing==3.3.1
pypdfium2==5.2.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.3
rarfile==4.2
requests==2.32.5
ruamel.yaml==0.19.1
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.1
termcolor==3.3.0
tifffile==2025.5.10
tqdm==4.67.1
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.2
visualdl==2.5.3
wcwidth==0.2.14
Werkzeug==3.1.4

# ============================================================
# Care Giver (Kelvin) - Additional Dependencies
# ============================================================

# Google Generative AI (Gemini) - for AI responses
google-generativeai>=0.3.0

# Google Cloud TTS/STT - for voice input/output (optional - has free fallbacks)
google-cloud-texttospeech>=2.14.0
google-cloud-speech>=2.21.0

# Free TTS/STT fallbacks (no credentials needed)
gTTS>=2.3.0
SpeechRecognition>=3.10.0

# Environment variable loading
python-dotenv>=1.0.0

# ROS2 camera interface (for MIPI camera)
# Note: rclpy is installed via ROS2, not pip
# Run: source /opt/tros/humble/setup.bash

# Audio dependencies (system packages, not pip)
# Install via: sudo apt install alsa-utils pulseaudio-utils
